+++ main.html
@@ -60,7 +60,7 @@
 		<div class="row">
 			<div id="waveform"></div>
 		</div>
-		<!-- JS SCRIPT IMPORTS  -->
+		<!-- JS SCRIPT IMPORTS -->
 		<script>
 			if (typeof module === 'object') {
 				window.module = module;
@@ -68,7 +68,7 @@
 			}
 		</script>
 		<!-- Main JS script and jQuery -->
-		<script src="scripts/vendor/jquery-3.1.1.min.js"></script>
+		<script src="scripts/vendor/jquery-3.1.1-min.js"></script>
 		<script src="data/keyCodes.js"></script>
 		<script src="scripts/main.js"></script>
 		<script src="scripts/vendor/sortable.js"></script>
